sudah

+++ app/Http/Controllers/JenisLaboratoriumController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Models\JenisLaboratorium;
 use App\Models\JenisLaboratoriumKomlis;
+use App\Models\Komli;
+use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\StoreJenisLaboratoriumRequest;
 use App\Http\Requests\UpdateJenisLaboratoriumRequest;
 
@@ -37,21 +39,25 @@ class JenisLaboratoriumController extends Controller
      */
     public function store(StoreJenisLaboratoriumRequest $request)
     {
-        $validatedData = $request->validate([
-            'jenis' => 'required',
-            'komli_id' => 'required'
-        ]);
-
-        $jenis = JenisLaboratorium::create($validatedData);
-
-        foreach ($request->komli_id as $key => $komli_id) {
-            JenisLaboratoriumKomlis::create([
-                'jenis_laboratorium_id' => $jenis->id,
-                'komli_id' => $komli_id
+        if (Auth::user()->hasRole('dinas')) {
+            $validatedData = $request->validate([
+                'jenis' => 'required',
+                'komli_id' => 'required'
             ]);
+    
+            $jenis = JenisLaboratorium::create($validatedData);
+    
+            foreach ($request->komli_id as $key => $komli_id) {
+                JenisLaboratoriumKomlis::create([
+                    'jenis_laboratorium_id' => $jenis->id,
+                    'komli_id' => $komli_id
+                ]);
+            }
+    
+            return redirect()->back();
+        }else{
+            abort(403);
         }
-
-        return redirect()->back();
 
     }
 
@@ -63,7 +69,17 @@ class JenisLaboratoriumController extends Controller
      */
     public function show(JenisLaboratorium $jenisLaboratorium)
     {
-        //
+        if(Auth::user()->hasRole('dinas')){
+            $komlis = getJenis($jenisLaboratorium->id);
+    
+            return view('admin.detail-lab',[
+                'jenis' => $jenisLaboratorium,
+                'komlis' => $komlis,
+                'option_komlis' => Komli::belumDipilihLab()
+            ]);
+        }else{
+            abort(403);
+        }
     }
 
     /**
@@ -86,7 +102,17 @@ class JenisLaboratoriumController extends Controller
      */
     public function update(UpdateJenisLaboratoriumRequest $request, JenisLaboratorium $jenisLaboratorium)
     {
-        //
+        if (Auth::user()->hasRole('dinas')) {
+            $validatedData = $request->validate([
+                'jenis' => 'required'
+            ]);
+    
+            $jenisLaboratorium->update($validatedData);
+    
+            return redirect()->back();
+        }else{
+            abort(403);
+        }
     }
 
     /**
@@ -97,10 +123,14 @@ class JenisLaboratoriumController extends Controller
      */
     public function destroy(JenisLaboratorium $jenisLaboratorium)
     {
-        foreach ($jenisLaboratorium->jenislaboratoriumkomli as $key => $jenis_komli) {
-            JenisLaboratoriumKomlis::destroy($jenis_komli->id);
+        if (Auth::user()->hasRole('dinas')) {
+            foreach ($jenisLaboratorium->jenislaboratoriumkomli as $key => $jenis_komli) {
+                JenisLaboratoriumKomlis::destroy($jenis_komli->id);
+            }
+            JenisLaboratorium::destroy($jenisLaboratorium->id);
+            return redirect()->back();
+        }else{
+            abort(403);
         }
-        JenisLaboratorium::destroy($jenisLaboratorium->id);
-        return redirect()->back();
     }
 }
